WebSockets/tests/submissions/Opera/interfaces/WebSocket/events/008.html

<!doctype html>
<title>WebSockets: listening for events with onerror</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  var run = false;
  function foo(e) {
    run = true;
    assert_equals(e.detail, 5)
  }
  ws.onerror = foo;
  var ev = document.createEvent('UIEvents');
  ev.initUIEvent('error', false, false, window, 5);
  ws.dispatchEvent(ev);
  assert_true(run);
});
</script>
